src/app/Classes/BinaryHeap/PriorityQueue.ts
import Node from './Node';

interface IPriorityQueueProps<T> {
  values: Node<T>[]
  enqueue: (value: T, priority: number) => PriorityQueue<T>
  bubbleUp: () => void
  dequeue: () => Node<T> | undefined
  sinkDown: () => void
}

export default class PriorityQueue<T> implements IPriorityQueueProps<T> {
  values: Node<T>[] = [];

  constructor() {
    this.values = [];
  }

  enqueue(value: T, priority: number): PriorityQueue<T> {
    let newNode = new Node<T>(value, priority);
    this.values.push(newNode);
    this.bubbleUp();
    return this;
  }

  bubbleUp() {
    let index = this.values.length - 1;
    const element = this.values[index];
    while (index > 0) {
      let parentIndex = Math.floor((index - 1) / 2);
      let parent = this.values[parentIndex];
      if (element.priority >= parent.priority) break;
      this.values[parentIndex] = element;
      this.values[index] = parent;
      index = parentIndex;
    }
  }

  dequeue(): Node<T> | undefined {
    const min = this.values[0];
    const end = this.values.pop();
    if (this.values.length > 0) {
      this.values[0] = end!;
      this.sinkDown();
    }
    return min;
  }

  sinkDown() {
    let index = 0;
    const length = this.values.length;
    const element = this.values[0];
    while (true) {
      let leftChildIdx = 2 * index + 1;
      let rightChildIdx = 2 * index + 2;
      let leftChild: Node<T>;
      let rightChild: Node<T>;
      let swap = null;

      if (leftChildIdx < length) {
        leftChild = this.values[leftChildIdx];
        if (leftChild.priority < element.priority) {
          swap = leftChildIdx;
        }
      }
      if (rightChildIdx < length) {
        rightChild = this.values[rightChildIdx];
        if (
          (swap === null && rightChild.priority < element.priority) ||
          (swap !== null && rightChild.priority < leftChild!.priority)
        ) {
          swap = rightChildIdx;
        }
      }
      if (swap === null) break;
      this.values[index] = this.values[swap];
      this.values[swap] = element;
      index = swap;
    }
  }
}





